fixed the receipt, delivery, and transfer flow

+++ src/main/java/whizware/whizware/dto/goods/GoodsRequest.java
@@ -2,11 +2,13 @@ package whizware.whizware.dto.goods;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class GoodsRequest {
 
     private String name;
-    private Long sellingPrice;
-    private Long purchasePrice;
+    private BigDecimal sellingPrice;
+    private BigDecimal purchasePrice;
     private String description;
 }
